clarity_backend/requirements.txt
django
twilio
gunicorn
python-decouple
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
psycopg2-binary
dj-database-url
channels[daphne]
channels-redis
redis